requirements.txt
alembic==1.14.0
babel==2.16.0
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
click==8.1.8
colorama==0.4.6
cssselect2==0.7.0
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
fpdf==1.7.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
meson==1.6.1
ninja==1.11.1.3
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
PyMuPDF==1.25.1
pyodbc==5.2.0
PyPDF2==3.0.1
pyphen==0.17.0
python-dotenv==1.0.1
reportlab==4.2.5
SQLAlchemy==2.0.36
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
weasyprint==63.1
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
